Fix for compile

+++ emby/PlatformSpecific/STM32xx_Baremetal/EmbyThreading/Thread.cc
@@ -1,13 +1,13 @@
 #include <EmbyThreading/Thread.hh>
 #include <EmbyThreading/Worker.hh>
-#include "EmbyDebug/assert.hh" s
+#include "EmbyDebug/assert.hh"
 namespace EmbyThreading
 {
 
 
     Thread::Thread( EmbyThreading::Worker* worker,
     			    char const* name,
-    			    uint32_t stackSize,
+    			    size_t stackSize,
                     Priority priority , bool start,uint32_t *stack): m_worker{worker}
 	{
